1: Renamed PolariBibleGlobals.py to PolariBible.py
