+++ PickleToLaTeX.py
@@ -12,7 +12,7 @@ Id = ""
 import pickle
 import re
 import sys
-import PolariBibleGlobals
+import PolariBible
 
 LaTeXHeader = r"""
 \documentclass[a4paper,10pt]{book}
